+++ ui/spec/data_explorer/reducers/queryConfigSpec.js
@@ -43,10 +43,13 @@ describe('Chronograf.Reducers.queryConfig', () => {
     })
 
     it('sets the db and rp', () => {
-      const newState = reducer(state, chooseNamespace(queryId, {
-        database: 'telegraf',
-        retentionPolicy: 'monitor',
-      }))
+      const newState = reducer(
+        state,
+        chooseNamespace(queryId, {
+          database: 'telegraf',
+          retentionPolicy: 'monitor',
+        })
+      )
 
       expect(newState[queryId].database).to.equal('telegraf')
       expect(newState[queryId].retentionPolicy).to.equal('monitor')
@@ -63,23 +66,33 @@ describe('Chronograf.Reducers.queryConfig', () => {
     let state
     beforeEach(() => {
       const one = reducer({}, fakeAddQueryAction('any', queryId))
-      const two = reducer(one, chooseNamespace(queryId, {
-        database: '_internal',
-        retentionPolicy: 'daily',
-      }))
+      const two = reducer(
+        one,
+        chooseNamespace(queryId, {
+          database: '_internal',
+          retentionPolicy: 'daily',
+        })
+      )
       const three = reducer(two, chooseMeasurement(queryId, 'disk'))
-      state = reducer(three, toggleField(queryId, {field: 'a great field', funcs: []}))
+      state = reducer(
+        three,
+        toggleField(queryId, {field: 'a great field', funcs: []})
+      )
     })
 
     describe('choosing a new namespace', () => {
-      it('clears out the old measurement and fields', () => { // what about tags?
+      it('clears out the old measurement and fields', () => {
+        // what about tags?
         expect(state[queryId].measurement).to.exist
         expect(state[queryId].fields.length).to.equal(1)
 
-        const newState = reducer(state, chooseNamespace(queryId, {
-          database: 'newdb',
-          retentionPolicy: 'newrp',
-        }))
+        const newState = reducer(
+          state,
+          chooseNamespace(queryId, {
+            database: 'newdb',
+            retentionPolicy: 'newrp',
+          })
+        )
 
         expect(newState[queryId].measurement).not.to.exist
         expect(newState[queryId].fields.length).to.equal(0)
@@ -87,13 +100,19 @@ describe('Chronograf.Reducers.queryConfig', () => {
     })
 
     describe('choosing a new measurement', () => {
-      it('leaves the namespace and clears out the old fields', () => { // what about tags?
+      it('leaves the namespace and clears out the old fields', () => {
+        // what about tags?
         expect(state[queryId].fields.length).to.equal(1)
 
-        const newState = reducer(state, chooseMeasurement(queryId, 'newmeasurement'))
+        const newState = reducer(
+          state,
+          chooseMeasurement(queryId, 'newmeasurement')
+        )
 
         expect(state[queryId].database).to.equal(newState[queryId].database)
-        expect(state[queryId].retentionPolicy).to.equal(newState[queryId].retentionPolicy)
+        expect(state[queryId].retentionPolicy).to.equal(
+          newState[queryId].retentionPolicy
+        )
         expect(newState[queryId].fields.length).to.equal(0)
       })
     })
@@ -103,10 +122,48 @@ describe('Chronograf.Reducers.queryConfig', () => {
         expect(state[queryId].fields.length).to.equal(1)
 
         const isKapacitorRule = true
-        const newState = reducer(state, toggleField(queryId, {field: 'a different field', funcs: []}, isKapacitorRule))
+        const newState = reducer(
+          state,
+          toggleField(
+            queryId,
+            {field: 'a different field', funcs: []},
+            isKapacitorRule
+          )
+        )
 
         expect(newState[queryId].fields.length).to.equal(1)
         expect(newState[queryId].fields[0].field).to.equal('a different field')
+      })
+    })
+
+    describe('TOGGLE_FIELDS', () => {
+      it('can toggle multiple fields', () => {
+        expect(state[queryId].fields.length).to.equal(1)
+
+        const newState = reducer(
+          state,
+          toggleField(queryId, {field: 'a different field', funcs: []})
+        )
+
+        expect(newState[queryId].fields.length).to.equal(2)
+        expect(newState[queryId].fields[1].field).to.equal('a different field')
+      })
+
+      it('applies a funcs to newly selected fields', () => {
+        expect(state[queryId].fields.length).to.equal(1)
+
+        const oneFieldOneFunc = reducer(
+          state,
+          applyFuncsToField(queryId, {field: 'a great field', funcs: ['func1']})
+        )
+
+        const newState = reducer(
+          oneFieldOneFunc,
+          toggleField(queryId, {field: 'a different field', funcs: []})
+        )
+
+        expect(newState[queryId].fields[1].funcs.length).to.equal(1)
+        expect(newState[queryId].fields[1].funcs[0]).to.equal('func1')
       })
     })
   })
@@ -192,7 +249,7 @@ describe('Chronograf.Reducers.queryConfig', () => {
       })
     })
 
-    it('creates a new entry if it\'s the first key', () => {
+    it("creates a new entry if it's the first key", () => {
       const initialState = {
         [queryId]: buildInitialState(queryId, {
           tags: {},
@@ -283,7 +340,9 @@ describe('Chronograf.Reducers.queryConfig', () => {
 
       const nextState = reducer(initialState, action)
 
-      expect(nextState[queryId].areTagsAccepted).to.equal(!initialState[queryId].areTagsAccepted)
+      expect(nextState[queryId].areTagsAccepted).to.equal(
+        !initialState[queryId].areTagsAccepted
+      )
     })
   })
 
@@ -314,7 +373,7 @@ describe('Chronograf.Reducers.queryConfig', () => {
     expect(nextState[queryId]).to.deep.equal(expected)
   })
 
-  it('updates a query\'s raw text', () => {
+  it("updates a query's raw text", () => {
     const initialState = {
       [queryId]: buildInitialState(queryId),
     }
@@ -326,7 +385,7 @@ describe('Chronograf.Reducers.queryConfig', () => {
     expect(nextState[queryId].rawText).to.equal('foo')
   })
 
-  it('updates a query\'s raw status', () => {
+  it("updates a query's raw status", () => {
     const initialState = {
       [queryId]: buildInitialState(queryId),
     }