+++ ui/src/side_nav/containers/SideNav.js
@@ -19,16 +19,12 @@ const SideNav = React.createClass({
     location: shape({
       pathname: string.isRequired,
     }).isRequired,
-    me: shape({
-      email: string,
-    }),
     isHidden: bool.isRequired,
     logoutLink: string,
   },
 
   render() {
     const {
-      me,
       params: {sourceID},
       location: {pathname: location},
       isHidden,
@@ -37,7 +33,7 @@ const SideNav = React.createClass({
 
     const sourcePrefix = `/sources/${sourceID}`
     const dataExplorerLink = `${sourcePrefix}/chronograf/data-explorer`
-    const showLogout = !!(me && me.name)
+    const showLogout = !!logoutLink
 
     return isHidden
       ? null
@@ -81,7 +77,7 @@ const SideNav = React.createClass({
           </NavBlock>
           {showLogout
             ? <NavBlock icon="user" className="sidebar__square-last">
-                <NavHeader link={logoutLink} title="Logout" />
+                <NavHeader useAnchor={true} link={logoutLink} title="Logout" />
               </NavBlock>
             : null}
         </NavBar>
@@ -89,10 +85,9 @@ const SideNav = React.createClass({
 })
 
 const mapStateToProps = ({
-  auth: {me, logoutLink},
+  auth: {logoutLink},
   app: {ephemeral: {inPresentationMode}},
 }) => ({
-  me,
   isHidden: inPresentationMode,
   logoutLink,
 })