+++ ui/src/utils/queryTransitions.js
@@ -16,31 +16,44 @@ export function chooseMeasurement(query, measurement) {
   })
 }
 
-export function toggleField(query, {field, funcs}, isKapacitorRule = false) {
+export const toggleField = (query, {field, funcs}, isKapacitorRule = false) => {
   const isSelected = query.fields.find(f => f.field === field)
   if (isSelected) {
     const nextFields = query.fields.filter(f => f.field !== field)
     if (!nextFields.length) {
-      const nextGroupBy = Object.assign({}, query.groupBy, {time: null})
-      return Object.assign({}, query, {
+      const nextGroupBy = {...query.groupBy, time: null}
+      return {
+        ...query,
         fields: nextFields,
         groupBy: nextGroupBy,
-      })
+      }
     }
 
-    return Object.assign({}, query, {
+    return {
+      ...query,
       fields: nextFields,
-    })
+    }
   }
 
   if (isKapacitorRule) {
-    return Object.assign({}, query, {
+    return {
+      ...query,
       fields: [{field, funcs}],
-    })
+    }
   }
-  return Object.assign({}, query, {
-    fields: query.fields.concat({field, funcs}),
-  })
+
+  let newFuncs = ['mean']
+  if (query.fields.length) {
+    newFuncs = query.fields.find(f => f.funcs).funcs
+  }
+
+  return {
+    ...query,
+    fields: query.fields.concat({
+      field,
+      funcs: newFuncs,
+    }),
+  }
 }
 
 export function groupByTime(query, time) {