+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 1. [#1530](https://github.com/influxdata/chronograf/pull/1530): Update query config field ordering to always match input query
 1. [#1535](https://github.com/influxdata/chronograf/pull/1535): Fix add field functions to existing Kapacitor rules
+1. [#1564](https://github.com/influxdata/chronograf/pull/1564): Fix regression of logout menu item functionality
 1. [#1562](https://github.com/influxdata/chronograf/pull/1562): Fix InfluxQL parsing with multiple tag values for a tag key
 
 ### Features
@@ -14,6 +15,7 @@
   1. [#1549](https://github.com/influxdata/chronograf/pull/1549): Reset graph zoom when a new time range is selected
   1. [#1544](https://github.com/influxdata/chronograf/pull/1544): Upgrade to new version of Influx Theme, remove excess stylesheets
   1. [#1567](https://github.com/influxdata/chronograf/pull/1567): Replace outline style User icon with solid style
+  1. [#1560](https://github.com/influxdata/chronograf/pull/1560): Apply mean to fields by default
   1. [#1561](https://github.com/influxdata/chronograf/pull/1561): Disable query save in dashboard editing if the query does not have a database, measurement, and field
 
 ## v1.3.1.0 [2017-05-22]